DMP-3963 - add media Linked case entries to addAudio, and make async
