+++ src/main/java/uk/gov/hmcts/darts/audio/service/AudioUploadService.java
@@ -8,6 +8,4 @@ public interface AudioUploadService {
     void addAudio(MultipartFile audioFileStream, AddAudioMetadataRequest addAudioMetadataRequest);
 
     void linkAudioToHearingInMetadata(AddAudioMetadataRequest addAudioMetadataRequest, MediaEntity mediaEntity);
-
-    void linkAudioToHearingByEvent(AddAudioMetadataRequest addAudioMetadataRequest, MediaEntity savedMedia);
 }
